Fixes validates with :uniqueness option.
